Create gbEater.py
This is the script, it is pretty simple!
Try to modify it but dont use it for bad use!
The default script will eat about 1.4GB of storage!
The script creates new txt files and writes the text in the:
file.write() line. 
It renames the files to be 0.txt, 1.txt and so along!